Combine modules `certificate` and `certificates`
Ref syslabcom/scrum#2142
